corrected notifications time

+++ chrome/android/java/src/org/chromium/chrome/browser/toolbar/ToolbarPhone.java
@@ -3142,7 +3142,7 @@ public class ToolbarPhone
     public void OnGetCurrentBalanceReport(String[] report) {}
 
     @Override
-    public void OnNotificationAdded(String id, int type, int timestamp,
+    public void OnNotificationAdded(String id, int type, long timestamp,
             String[] args) {
         if (mBraveRewardsNativeWorker != null) {
             mBraveRewardsNativeWorker.GetAllNotifications();
@@ -3171,7 +3171,7 @@ public class ToolbarPhone
     }
 
     @Override
-    public void OnGetLatestNotification(String id, int type, int timestamp,
+    public void OnGetLatestNotification(String id, int type, long timestamp,
             String[] args) {}
 
     @Override
